+++ engines/nunjucks.js
@@ -7,7 +7,7 @@ export default class Denjucks extends TemplateEngine {
   constructor(site, options) {
     super(site, options);
 
-    const loader = new nunjucks.FileSystemLoader(this.includes);
+    const loader = new nunjucks.FileSystemLoader(site.src("_includes"));
     this.engine = new nunjucks.Environment(loader, options);
 
     // Update cache