+++ plugins/modules.js
@@ -3,7 +3,7 @@ import loader from "../loaders/module.js";
 
 export default function () {
   return (site) => {
-    site.engine([".tmpl.js", ".tmpl.ts"], new Module(site));
+    site.loadPages([".tmpl.js", ".tmpl.ts"], loader, new Module(site));
     site.loadData([".js", ".ts"], loader);
   };
 }