+++ plugins/nunjucks.js
@@ -1,4 +1,5 @@
 import NunjucksEngine from "../engines/nunjucks.js";
+import loader from "../loaders/text.js";
 import { merge } from "../utils.js";
 
 // Default options
@@ -13,7 +14,7 @@ export default function (userOptions) {
   return (site) => {
     const nunjucksEngine = new NunjucksEngine(site, options.options);
 
-    site.engine(options.extensions, nunjucksEngine);
+    site.loadPages(options.extensions, loader, nunjucksEngine);
     site.filter("njk", filter, true);
 
     function filter(string, data = {}) {